Streamlit/requirements.txt
JPype1==1.5.0
nltk==3.8.1
numpy==1.26.4
streamlit==1.34.0
tensorflow==2.15.1
tensorflow==2.16.1
tensorflow_intel==2.15.1
tensorflow_intel==2.16.1
